Reinstate `transform` in ParallelClustering as not yet implemented
The removal of `transform` by ccdf704 triggers errors when MapperPipeline objects are instantiated, because scikit-learn requires that `transform` methods be implemented by all intermediate transformers of a Pipeline object.

Reinstating it as a not yet implemented method has the following consequences:
- MapperPipeline objects can be instantiated with no errors
- The behaviour upon calling `fit_transform` is the correct one as the implemented `fit_transform` method is used.
- calling `fit` on a MapperPipeline object and then `transform` will fail as a NotImplementedError is raised. 

The last point is likely a good thing as we avoid suggesting to the user that assignments of new points to clusters seen in `fit` are available.