Document Nerve class, apply adapt_fit_transform_docs, hide attributes
But Nerve needs to be refactored to remove fit and transform in favour of fit_transform. It is misleading to have a transform method which completely ignores X and only uses the X seen in fit. Nerve needs to be turned into a pure fit-transformer (implement only fit_transform), but this means the MapperPipeline output by make_mapper_pipeline needs to be augmented with a finall 'passthrough' step since otherwise the scikit-learn API would complain about the last step not implementing a fit method.