Remove check_images and additions to test_common
These are valuable and thank you @wreise, but in the interest of pushing at least check_point_clouds, I am opting for a more ad-hoc implementation of the latter and for the temporary removal of check_images. As we also did not have much time to discuss the additions to test_common, I will also revert that file to the state on master for now. It can be reintroduced in a later PR together with check_images