Increase generality of ripser function when metric == 'precomputed'
Sklearn's pairwise_distances is no longer called when metric='precomputed'.

This is because pairwise_distances checks for positivity and finite input in that case, but this is unnecessarily restrictive for ripser. 

pairwise_distances also checks for square input so this is no longer checked in the case metric='precomputed'.  But the ripser function is not meant for use in isolation. It is meant to be called by a VietorisRipsPersistence instance, and input validation can be offloaded to that class. In particular, 9bab872352cbbfd046425163c589c94e9ad37628 introduces checks for square input.