Allow more general input to ripser when metric='precomputed', improve/refactor check_point_clouds and add tests (#386)
* Remove boolean mask on infinite values in VietorisRipsPersistence

* Refactoring of check_point_clouds

- Rename distance_matrix kwarg to distance_matrices
- Add checks for square input when distance_matrices is set to True
- More comprehensive/friendly error messages and warnings

* Increase generality of ripser function when metric == 'precomputed'

Sklearn's pairwise_distances is no longer called when metric='precomputed'.

This is because pairwise_distances checks for positivity and finite input in that case, but this is unnecessarily restrictive for ripser. 

pairwise_distances also checks for square input so this is no longer checked in the case metric='precomputed'.  But the ripser function is not meant for use in isolation. It is meant to be called by a VietorisRipsPersistence instance, and input validation can be offloaded to that class. In particular, 9bab872352cbbfd046425163c589c94e9ad37628 introduces checks for square input.

* Add tests for check_point_clouds

* Refresh docstrings for VietorisRipsPersistence

- Mention "adjacency matrices of weighted undirected graphs" as well as distance matrices
- Mention compatibility with filtration in docstrings for fit and transform