Fixes error in _validate_clusterer
Function for checking if clusterer uses both precomputed affinity and metric (example:sklearn.cluster.AgglomerativeClustering) previously resulted in an error even when not using precomputed metric. This was because the list comprehension to check if there was a precomputed metric parameter resulted in a boolean list instead of a list of parameters set to "precomputed". This small change fixes that bug and ensures the expected behaviour.